+++ app/src/test/java/com/harmony/harmonyAndroid/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.harmony
+package com.harmony.harmonyAndroid
 
 import org.junit.Test
